Refactor `Option` into `Some` and `None` classes
This way we don't have to check `some?` and `none?`; we know what to do
by the type of the object.

Sandi Metz would be proud.

Note that the methods listed below had to be removed. They are the
functions on Rust's `Option` type that required a mutable reference to
`self` because they might leave a different variant in the `Option`'s
place. With the previous implementation, we could update `@variant` to
do that but now that's no longer possible.
- `get_or_insert`
- `get_or_insert_with`
- `replace`
- `take`