Don't use prelude in tests
Ideally the test suite should pass with or without the prelude.
Unfortunately, once the prelude is loaded it's non-trivial to unload it.
So I've rewritten the tests to not depend on the prelude. I'll take a
crack at writing a test just for the prelude later. Or I'll figure out
some fancy way of using `setup` / `teardown` to load & unload the
prelude.

This commit is best viewed with the `-w` / `--ignore-all-space` option
due to the shifting of the `match()` context